requirements.txt
bottle==0.12.18
cycler==0.10.0
Jinja2==2.11.2
joblib==0.15.1
kiwisolver==1.2.0
llvmlite==0.32.1
MarkupSafe==1.1.1
matplotlib==3.2.2
numba==0.49.1
numpy==1.18.4
Pillow==7.1.2
pyparsing==2.4.7
python-dateutil==2.8.1
scipy==1.4.1
scikit-learn==0.19.0
six==1.15.0
threadpoolctl==2.1.0
